Added: Roles

+++ routes/register.router.js
@@ -47,43 +47,72 @@ registerRouter.get("/", async (req, res) => {
 })
 registerRouter.post("/", upload.single("idcard"), async (req, res) => {
     try {
-        let { name, email, password, section, course, rollno, school } = req.body;
+
+        let { name, email, password, section, course, rollno, school, role } = req.body;
         const user = await RegisterModel.findOne({ email });
-        const pendinguser = await PendingUserModel.findOne({ email })
-        if (!user || pendinguser) {
-            return res.status(400).json({ message: "User already exists." })
+        const pendinguser = await PendingUserModel.findOne({ email })     //users who are not approved are pending users
+        if (user || pendinguser) {
+            return res.status(400).json({ msg: "User already exists." })
         } else {
             console.log(req.file, "req.file")
             const token = crypto.randomBytes(16).toString('hex');
 
             const hashed = await bcrypt.hash(password, 2);
-            const data = { name, email, password: hashed, section, course, rollno, school, token };
-            console.log(data)
+            let data;
+            if (role == "student") {
+                data = { name, email, password: hashed, section, course, rollno, school, token, role };
+            } else if (role == "faculty") {
+                data = { name, email, password: hashed, token, role };
+            }
+            console.log(data, "data")
             await PendingUserModel.create(data)
 
-            let mailOptions = {
-                from: process.env.mail_admin,
-                to: process.env.mail_admin,
-                subject: 'New Registration',
-                text: `
+            let mail_body;
+            let mail_html;
+            if (role == "student") {
+                mail_body = `
               Name: ${req.body.name}
               School: ${req.body.school}
               Course: ${req.body.course}
               Section: ${req.body.section}
               Rollno: ${req.body.rollno}
               Email: ${req.body.email}
-            `,
-                html: `
-              <p>New user registration:</p>
-              <p>Name: ${req.body.name}</p>
-              <p>School: ${req.body.school}</p>
-              <p>Course: ${req.body.course}</p>
-              <p>Section: ${req.body.section}</p>
-              <p>Rollno: ${req.body.rollno}</p>
-              <p>Email: ${req.body.email}</p>
-              <button><a href="https://conference-pkoe.onrender.com/register/verify?token=${encodeURIComponent(token)}">Click here to verify</a></button><br/>
-              <button><a href="https://conference-pkoe.onrender.com/register/decline?token=${encodeURIComponent(token)}">Click here to decline</a></button>
-            `,
+              Role: ${req.body.role}
+            `
+                mail_html = `
+            <p>New student registration:</p>
+            <p>Name: ${req.body.name}</p>
+            <p>School: ${req.body.school}</p>
+            <p>Course: ${req.body.course}</p>
+            <p>Section: ${req.body.section}</p>
+            <p>Rollno: ${req.body.rollno}</p>
+            <p>Email: ${req.body.email}</p>
+            <p>Role: ${req.body.role}</p>
+            <button><a href="https://conference-pkoe.onrender.com/register/verify?token=${encodeURIComponent(token)}">Click here to verify</a></button><br/>
+            <button><a href="https://conference-pkoe.onrender.com/register/decline?token=${encodeURIComponent(token)}">Click here to decline</a></button>
+          `
+            } else if (role == "faculty") {
+                mail_body = `
+              Name: ${req.body.name}
+              Email: ${req.body.email}
+              Role: ${req.body.role}
+            `
+                mail_html = `
+            <p>New faculty registration:</p>
+            <p>Name: ${req.body.name}</p>
+            <p>Email: ${req.body.email}</p>
+            <p>Role: ${req.body.role}</p>
+            <button><a href="http://localhost:8080/register/verify?token=${encodeURIComponent(token)}">Click here to verify</a></button><br/>
+            <button><a href="http://localhost:8080/register/decline?token=${encodeURIComponent(token)}">Click here to decline</a></button>
+          `
+            }
+
+            let mailOptions = {
+                from: process.env.mail_admin,
+                to: process.env.mail_admin,
+                subject: 'New Registration',
+                text: mail_body,
+                html: mail_html,
                 attachments: [
                     {
                         filename: req.file.originalname,
@@ -98,36 +127,45 @@ registerRouter.post("/", upload.single("idcard"), async (req, res) => {
                 }
                 console.log('Message sent: %s', info.messageId);
             })
-            res.status(200).send({ "msg": "Send for verification" })
+            res.status(200).send({ "msg": "Send for verification", ok: true })
         }
     } catch (err) {
         console.log(err);
-
+        res.send({ msg: "Someting went wrong", ok: false })
     }
 })
 
 registerRouter.get("/verify", async (req, res) => {
     try {
         const token = req.query.token;
-        const { name, email, password, section, course, rollno, school } = await PendingUserModel.findOne({ token })
+        const pending_user =  await PendingUserModel.findOne({ token });
+        if(!pending_user){
+            res.send("User Already Exists");
+            return;
+        }
+        const { name, email, password, section, course, rollno, school, role } = pending_user;
 
         const user = await RegisterModel.findOne({ email })
         if (user) {
             await PendingUserModel.deleteOne({ token })
             res.status(409).send({ msg: "User already exists" })
         } else {
+            const handle_token = crypto.randomBytes(2).toString('hex');
             await RegisterModel.create({
-                name: name,
-                email: email,
-                password: password,
-                school: school,
-                course: course,
-                section: section,
-                rollno: rollno,
-                handle:name
+                name,
+                email,
+                password,
+                school,
+                course,
+                section,
+                rollno,
+                handle: `${name}${handle_token}`,
+                role
             });
-            const user = await RegisterModel.findOne({ email });
-            console.log(user)
+            await PendingUserModel.deleteOne({ token })
+
+            // const user = await RegisterModel.findOne({ email });
+            // console.log(user)
 
             transporter.sendMail({
                 to: email,
@@ -149,13 +187,13 @@ Founder and CEO`
             })
 
                 .then(() => {
-                    console.log("mail send successfully!")
+                    console.log("mail send successfully!");
+
                 })
                 .catch((err) => {
                     console.log("Error while sendig mail")
                     console.log(err)
                 })
-            await PendingUserModel.deleteOne({ token })
             res.status(201).send("verified, Account Created")
         }
     } catch (error) {
@@ -168,8 +206,12 @@ registerRouter.get("/decline", async (req, res) => {
     try {
         const token = req.query.token;
         const user = await PendingUserModel.findOne({ token });
+        if (!user) {
+            res.send("User already exits");
+            return
+        }
         console.log(user, "user")
-        let { name, email, password, section, course, rollno, school } = user;
+        let { name, email, password, section, course, rollno, school , role} = user;
 
         transporter.sendMail({
             to: email,
